% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vetor_binario.R
\name{modelos_possiveis}
\alias{modelos_possiveis}
\title{Adiciona e remove um elemento por vez}
\usage{
modelos_possiveis(vet)
}
\arguments{
\item{vet}{vetor binário.}
}
\value{
um data frame com o conjunto de combinações possíveis.
}
\description{
Adiciona e remove um elemento por vez
}
\examples{
modelos_possiveis(c(1,1,0))

}
